Rename `ProjectTaskUtil` -> `ProjectConfigurations`
Also divide the class into two
